requirements.txt
fastapi
openai
require
python-docx
uvicorn
# ts_gene
python-multipart
langchain
langchain-core
langchain-openai
langchain-community
langchain-chroma
langchain-groq
python-dotenv   
sentence-transformers                    